1차 병합 및 코드 수정 (#33)
* reformat code

* test: test

* update: config 내 파일 정리 및 오탈자 수정

* update: 컨트롤러 수정

* update: 엔티티, 리스폰스, 리퀘스트 수정

* update: repository 어노테이션 추가

* update: user.css layout에 추가

* update: 불필요한 주석 제거

* update: css체크

+++ src/main/java/carrotmoa/carrotmoa/model/request/CreateAccommodationRequest.java
@@ -3,15 +3,18 @@ package carrotmoa.carrotmoa.model.request;
 import carrotmoa.carrotmoa.entity.Accommodation;
 import carrotmoa.carrotmoa.entity.AccommodationSpace;
 import carrotmoa.carrotmoa.entity.Post;
-import jakarta.validation.constraints.*;
-import lombok.*;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.multipart.MultipartFile;
-
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.multipart.MultipartFile;
 
 @Getter
 @Setter
@@ -65,36 +68,36 @@ public class CreateAccommodationRequest {
 
     public Accommodation toAccommodationEntity() {
         return Accommodation.builder()
-                .totalArea(totalArea)
-                .roadAddress(roadAddress)
-                .lotAddress(lotAddress)
-                .detailAddress(detailAddress)
-                .floor(floor)
-                .totalFloor(totalFloor)
-                .price(price)
-                .transportationInfo(transportationInfo)
-                .build();
+            .totalArea(totalArea)
+            .roadAddress(roadAddress)
+            .lotAddress(lotAddress)
+            .detailAddress(detailAddress)
+            .floor(floor)
+            .totalFloor(totalFloor)
+            .price(price)
+            .transportationInfo(transportationInfo)
+            .build();
     }
 
     public Post toPostEntity() {
         return Post.builder()
-                .serviceId(SERVICE_ID)
-                .userId(userId)
-                .title(title)
-                .content(content)
-                .build();
+            .serviceId(SERVICE_ID)
+            .userId(userId)
+            .title(title)
+            .content(content)
+            .build();
     }
 
     public List<AccommodationSpace> toAccommodationSpaceEntities() {
         return accommodationSpaces.stream()
-                .map(accommodationSpaceRequest -> {
-                    AccommodationSpace accommodationSpace = new AccommodationSpace();
-                    accommodationSpace.setAccommodationId(accommodationSpaceRequest.getAccommodationId());
-                    accommodationSpace.setSpaceId(accommodationSpaceRequest.getSpaceId());
-                    accommodationSpace.setCount(accommodationSpaceRequest.getCount());
-                    return accommodationSpace;
-                })
-                .collect(Collectors.toList());
+            .map(accommodationSpaceRequest -> {
+                AccommodationSpace accommodationSpace = new AccommodationSpace();
+                accommodationSpace.setAccommodationId(accommodationSpaceRequest.getAccommodationId());
+                accommodationSpace.setSpaceId(accommodationSpaceRequest.getSpaceId());
+                accommodationSpace.setCount(accommodationSpaceRequest.getCount());
+                return accommodationSpace;
+            })
+            .collect(Collectors.toList());
     }
 
     // 공간 초기화 메서드
@@ -102,7 +105,7 @@ public class CreateAccommodationRequest {
         accommodationSpaces.clear(); // 기존 공간 초기화
         for (int i = 0; i < spaceCount; i++) {
             AccommodationSpaceRequest accommodationSpaceRequest = new AccommodationSpaceRequest();
-            accommodationSpaceRequest.setSpaceId((long) (i + 1));
+            accommodationSpaceRequest.setSpaceId(i + 1);
             accommodationSpaceRequest.setCount(0);
             accommodationSpaces.add(accommodationSpaceRequest);
         }
